Set clantag on options loaded
There is no guarantee that the clientprefs, and by extension gokz
options, have been loaded when OnClientPutInServer is called. This can
lead to situations where the clients load into the game server and get
the VNL tag on the scoreboard, instead of the mode they actually have
loaded. This has been seen when there is more load on the SQL server,
for example during map changes with many players.

So to fix that issue, instead of setting the initial clan tag for the
user when joining the server, we wait until the options have been
loaded.