Rearrange order of rank tresholds
The order of the modes go VNL, SKZ, KZT. The previous order of the rank
points meant that in KZT you needed least points to reach high rank, and
in VNL most. This means that almost everyone is a really high rank in
KZT, because its too easy to reach.

This change moves the order so that VNL require least points, and KZT
most.