Add ErrorBoxFixer plugin from KZTimer
There are models that are used in some maps that are not packaged with
the map itself. This means that if we want to be able to show them
properly for the client, we need to have the client download the models.
This problem has not been seen that often before, since if a player has
played on a KZTimer server before, he would have the models locally
already.

The code, models and materials here have been taken from the KZTimer
plugin, with some small modifications to the code to make it fit into the
GOKZ project.