Prevent SQL errors due to ONLY_FULL_GROUP_BY
When ONLY_FULL_GROUP_BY is set on the SQL server, the queries before this
change failed with an error. The ONLY_FULL_GROUP_BY mode is set by
default on new MySQL installations at least, so we could assume that
this will happen more and more as time goes by.

This fix adds missing columns to the group bys that are in the inner
queries. It also removes the group by from the outer most one, as we
have already grouped the results earlier, and there should be no need to
do it again.

As a disclaimer though, I am not a SQL wizard, and there might be some
edge-cases where this could break..? Maybe?