Show warning on CP in non-cp area
Adds a message that will show in case a player has the timer off and
creates a CP in an area that is marked to have CPs disabled. If the
timer is running, the player will not be able to create a CP.

It adds the message so that the user can test in what places its
possible to create CPs without having the timer on. This reduces
confusion and makes the routing process easier.