Add warning message when starting a non-global run
In case the map is the wrong version, or the server is not updated the
user might not realize that the run will not be saved in the global api.
To avoid the confusion, we show them an error message.

To prevent spam of the message, we first check if the server is intended
to be global, by verifying that the GlobalApi plugin has a key set.
Further if the user already has a run going when resetting the timer,
they will not get the message again. If they stop the timer and start it
again, it will show the message again.