Add gokz_warn_for_non_global_map cvar
This allows server owners to disable the warnings for non-global maps.
This is useful in cases where you host a lot of non-global maps on your
server. The default value is to warn users about non-global times, as
most servers that have the GlobalAPI Key set mainly host global maps.