Add additional resources from KZTimer
These were previously skipped in
d2d39ff751ab336a462532d22020244ed500b762, as it seemed as if they were
not required. However, it has now been seen that these files are used in
maps also, and are thus required to be shipped. An example map is
xc_nephilim, which will show error boxes if the files are not included
for the start timer (switch files).

In KZTimer they were a part of the base plugin, and not the ErrorBoxFixer,
but in GOKZ it makes more sense to have them all in the same place.